Moved module mock_client into mock
