+++ modules/src/ics26_routing/handler.rs
@@ -76,8 +76,8 @@ mod tests {
     use crate::ics26_routing::handler::dispatch;
     use crate::ics26_routing::msgs::ICS26Envelope;
     use crate::mock::context::MockContext;
-    use crate::mock_client::header::MockHeader;
-    use crate::mock_client::state::{MockClientState, MockConsensusState};
+    use crate::mock::header::MockHeader;
+    use crate::mock::client_state::{MockClientState, MockConsensusState};
     use crate::test_utils::get_dummy_account_id;
     use crate::Height;
