+++ modules/src/mock/mod.rs
@@ -1,4 +1,7 @@
-//! Implementation of mocks for context and host chain.
+//! Implementation of mocks for context, host chain, and client.
 
 pub mod context;
 pub mod host;
+pub mod header;
+pub mod client_def;
+pub mod client_state;
