--- modules/src/mock_client/mod.rs
@@ -1,5 +1,0 @@
-//! Mock client implementation, used for testing
-
-pub mod client_def;
-pub mod header;
-pub mod state;
